Support multiple servers in preference order
This relies on the the Net::LDAP behaviour of handling multiple server names. 
"HOST may also be a reference to an array of hosts, host-port pairs or URIs to try. Each will be tried in order until a connection is made. Only when all have failed will the result of undef be returned"